requirements.txt
streamlit
openai
gspread
oauth2client
Pillow
beautifulsoup4
